[WIP] MKL repos download for mac and windows + update to MKL(-DNN) 0.12 (#17396)
* Fix MKL build on MacOS

* Fix invalid return type, user_memory_->set_data_handle(buf) returns void.

* Fix type mismatch

* Implement downloading and linking of MKL repo on mac and windows

* Rework the way MKL libs are binded to avoid warnings and simplify code

* Switch to matching version of MKL-DNN

* Add mirror links

* Add renamed license file

* Add -fopenmp flag for linux mkl build as Slice op kernel is optimized using OpenMP

* Use clean_dep in the entire file for consistency

* Fix build. Constants are declrared as static in header outside the class scope now

* Reworked fix for `-openmp` flag using the suggestion from @gunan
Original version here https://github.com/gunan/tensorflow/commit/b4d39c5226e8bbafe084b1dfc35f4644f2d906c2

+++ tensorflow/core/kernels/mkl_fused_batch_norm_op.cc
@@ -933,7 +933,7 @@ class MklFusedBatchNormOp : public OpKernel {
   bool is_training_;
   T* mean_values_;
   T* variance_values_;
-  size_t depth_;  // batch normalization is done for per channel.
+  int depth_;  // batch normalization is done for per channel.
 
   void ExtractParams(OpKernelContext* context) {
     const Tensor& input = MklGetInput(context, 0);
